feat(controller): package 구조 수정
