+++ jwp-basic-gradle/src/main/java/next/api/auth/controller/LoginController.java
@@ -1,4 +1,4 @@
-package next.api.auth;
+package next.api.auth.controller;
 
 import core.web.ModelAndView;
 import javax.servlet.http.HttpServletRequest;
