+++ jwp-basic-gradle/src/main/java/next/api/qna/controller/ListQuestionApiController.java
@@ -1,22 +1,20 @@
-package next.api.qna;
+package next.api.qna.controller;
 
 import core.web.ModelAndView;
-
-import java.util.List;
-
-import next.common.controller.AbstractController;
 import next.api.qna.dao.QuestionDao;
 import next.api.qna.model.Question;
-import next.common.view.JspView;
+import next.common.controller.AbstractController;
+import next.common.view.JsonView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.sql.SQLException;
+import java.util.List;
 
-public class ListQuestionController extends AbstractController {
-    private static final Logger log = LoggerFactory.getLogger(ListQuestionController.class);
+public class ListQuestionApiController extends AbstractController {
+    private static final Logger log = LoggerFactory.getLogger(ListQuestionApiController.class);
     private final QuestionDao questionDao = QuestionDao.getInstance();
 
     @Override
@@ -29,6 +27,6 @@ public class ListQuestionController extends AbstractController {
             log.error(e.toString());
         }
 
-        return new ModelAndView(new JspView("/qna/list.jsp")).addModel("questions", questions);
+        return new ModelAndView(new JsonView()).addModel("questions", questions);
     }
 }
