+++ jwp-basic-gradle/src/main/java/next/api/user/controller/UpdateUserController.java
@@ -1,6 +1,10 @@
-package next.api.user;
+package next.api.user.controller;
 
 import core.web.ModelAndView;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
 import next.common.controller.AbstractController;
 import next.api.user.dao.UserDao;
 import next.api.user.model.User;
@@ -8,12 +12,10 @@ import next.common.view.JspView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.sql.SQLException;
 
-public class CreateUserController extends AbstractController {
-    private static final Logger log = LoggerFactory.getLogger(CreateUserController.class);
+public class UpdateUserController extends AbstractController {
+    private static final Logger log = LoggerFactory.getLogger(UpdateUserController.class);
     private final UserDao userDao = UserDao.getInstance();
 
     @Override
@@ -22,15 +24,16 @@ public class CreateUserController extends AbstractController {
                 request.getParameter("userId"),
                 request.getParameter("password"),
                 request.getParameter("name"),
-                request.getParameter("email")
-        );
-
+                request.getParameter("email"));
+        log.debug("updateUser : {}", user);
         try {
-            userDao.insert(user);
+            userDao.update(user);
         } catch (SQLException e) {
-            log.error(e.getMessage());
+            log.error(e.toString());
         }
 
+        HttpSession session = request.getSession();
+        session.setAttribute("user", user);
         return new ModelAndView(new JspView("redirect:/user/list"));
     }
 }
